Separated skin manager from PlayerObject
