+++ Assets/Prefabs/Player/PlayerSkinManager.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: cf361fba945c46dba9b470e5c3e67439
+timeCreated: 1745238166
